backend-deploy/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
mongo-motor==3.3.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
bcrypt==4.0.1
pydantic==2.5.0